Add VRR support to xwayland
This commit adds support for Variable Refresh Rate (VRR) to the xwayland component of the window manager. The 'misc' section in the wm.nix file has been modified to include the 'vrr' option with a value of 1, enabling VRR functionality.

This change allows for smoother and more responsive graphics rendering on displays that support VRR, resulting in an improved user experience.